+++ Code/bot/__init__.py
@@ -4,4 +4,4 @@ from discord.ext import commands
 intents = discord.Intents.all()
 bot = commands.Bot(command_prefix='!', intents=intents, help_command=None)
 
-from .main_bot_core import main
+from .main_bot_core import bot_start
