+++ Code/bot/main_bot_core.py
@@ -4,14 +4,14 @@ import discord.ext.commands
 from bot import bot
 from bot.help_command import help
 from bot.player_control import player_add, player_rm
-from db_work import 123
 from discord.ext import commands
 
-TOKEN_PATH: str = "secrets/token"
+from Code.main import settings_manager
+
 
 @bot.command()
 async def ping(ctx: discord.ext.commands.Context):
     await ctx.send(f"Pong in ({round(bot.latency * 1000)}ms)")
 
-async def main():
-    await bot.start(123(TOKEN_PATH).load_data())
+async def bot_start():
+    await bot.start(await settings_manager.get_setting("token"))
