+++ Code/html/pages/__init__.py
@@ -1,4 +1,5 @@
 from html.pages.about import render_about_main_page
 from html.pages.index import render_index_main_page
 from html.pages.players import render_players_main_page
-from html.pages.settings.main import render_settings_main_page
+
+from Code.html.pages.settings.settings_main import render_settings_main_page
